+++ frontend/src/app/interfaces/user.ts
@@ -1,0 +1,8 @@
+export interface User {
+  id: number;
+  login: string;
+  password: string;
+  email: string;
+  createdAt: string;
+  updatedAt: string;
+}