+++ frontend/src/app/services/exemple.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ExempleService } from './exemple.service';
+
+describe('ExempleService', () => {
+  let service: ExempleService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ExempleService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});