+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v1.1 - Exemples
 
-* Ajout d'exemples
+* Réécriture de la "Loader Directive"
+* Ajout de [ngx-sweetalert2](https://www.npmjs.com/package/@sweetalert2/ngx-sweetalert2)
+* Ajout de [Angular - Datatables](http://l-lin.github.io/angular-datatables/#/welcome)
+* Ajout de la possibilité d'utiliser les requêtes Swagger avec le CORS
+* Ajout d'exemples pour le backend et le frontend
 * Mise à jour des dépendances
 
 ## v1.0 - Initialisation